feat: add translator

+++ p2k2_converter/p2k2/classes/p2k2.py
@@ -43,7 +43,7 @@ class Machining:
             "type": "Attribute",
         }
     )
-    offset: Optional[int] = field(
+    offset: Optional[float] = field(
         default=None,
         metadata={
             "name": "OFFSET",
